web_scraping/requirements.txt
pandas==2.1.2
numpy==1.26.1
beautifulsoup4==4.12.2
requests==2.31.0
coverage==7.3.2
lxml==4.9.3
psycopg2-binary==2.9.9